requirements.txt
numpy==1.24.2
opencv_python==4.11.0.86
Pillow==11.1.0
PyQt5==5.15.11
PyQt5_sip==12.17.0
PyYAML==6.0.2
tqdm==4.67.1
pyinstaller